17367894/emf-agent-main/requirements.txt
langchain
langchain-community
langsmith
langchain_ollama
langgraph
tqdm
pandas
matplotlib
fastapi
mcp
langchain-mcp-adapters
langchain-openai
python-dotenv